Shadows: PCF receiver path, selectable via ShadowLayer.algorithm
Default cascade-shadow receiver is now 9-tap rotated PCF; MSM stays
available behind ShadowAlgorithm.MSM for IEEE-FP-strict GPUs. Skip
anisotropic filtering on render targets - Adreno applies it to the
RGBA32F cascade moments and noisifies the receiver path.

+++ worldwind/src/androidMain/kotlin/earth/worldwind/util/kgl/AndroidKgl.kt
@@ -322,7 +322,7 @@ class AndroidKgl : Kgl {
         }
     }
 
-    override fun fenceSync() = KglSync(GLES30.glFenceSync(GL_SYNC_GPU_COMMANDS_COMPLETE, 0).toLong())
+    override fun fenceSync() = KglSync(GLES30.glFenceSync(GL_SYNC_GPU_COMMANDS_COMPLETE, 0))
 
     override fun isSyncSignalled(sync: KglSync): Boolean {
         if (!sync.isValid()) return false
